# Conflicts:
#	src/ContactsArray.java
#	src/contactsManagerTest.java

+++ src/contactsManagerTest.java
@@ -11,13 +11,13 @@ import java.io.IOException;
 
 public class contactsManagerTest {
 
-    public static Contacts[] contactNew = ContactsArray.findAll();
+    public static Contacts[] contactNew = ContactsArray.findAll(); // calling array object of contacts
 
-    public static void main(String[] args) throws IOException {
+    public static void main(String[] args) { // main method
         openMenu();
     }
 
-    private static void openMenu(){
+    private static void openMenu(){ // menu items
         Scanner input = new Scanner(System.in);
         int userInput;
 
@@ -34,7 +34,7 @@ public class contactsManagerTest {
 
             switch(userInput){
                 case 1:
-                    contactNew();
+                    showContact();
                     break;
                 case 2:
                     addContact();
@@ -49,23 +49,23 @@ public class contactsManagerTest {
                     System.out.println("Thank you and please come back again!");
                     System.exit(0);
             }
-
             System.out.println("\nEnter 'y' to continue.");
-            if(input.next().toLowerCase().startsWith("y")){
+            if(input.next().toLowerCase().startsWith("y")){  // if user decides to continue, go back to main menu, if not exit.
                 openMenu();
             } else
                 System.exit(0);
-
-
         }while(true);
-
     }
 
-    private static void contactNew() {
-        System.out.printf("%-23s|%15s\n", "Name", "Phone number");
-        System.out.printf("%s\n", "---------------------------------------");
-        for(Contacts contact: contactNew)
-            System.out.printf("%-20s   |   %s\n", contact.getFullName(), contact.getPhoneNumber());
+    private static void showContact(){
+        try {
+            System.out.printf("%-23s|%15s\n", "Name", "Phone number"); // displaying tittle menu
+            System.out.printf("%s\n", "---------------------------------------");
+            for (Contacts contact : contactNew)
+                System.out.printf("%-20s   |   %s\n", contact.getFullName(), contact.getPhoneNumber());
+        } catch (NoSuchFieldError  NosuchFile){
+            System.err.println("Exception " + NosuchFile);
+        }
     }
 
     public static void addContact() {
@@ -79,14 +79,37 @@ public class contactsManagerTest {
                 String userNewFirst = input.next();
                 System.out.println("Please enter new Last name:");
                 String userNewLast = input.next();
-                System.out.println("Please enter new phone number: ");
-                String userNewNumber = input.next();
-                List<String> lines = Files.readAllLines(path, StandardCharsets.UTF_8);
-                int position = 3;
-                String extraline = String.format("%20s(\"%s %s\",\"%s\"),", "new Contacts", userNewFirst, userNewLast, userNewNumber);
-                System.out.println(extraline);
-                lines.add(position, extraline);
-                Files.write(path, lines, StandardCharsets.UTF_8);
+                String fullContactName = String.format("%s %s", userNewFirst, userNewLast);
+
+                for(Contacts contact: contactNew) {
+                    if (contact.getFullName().equalsIgnoreCase(fullContactName)) {
+                        System.out.printf("\n%s %s\n\n", contact.getFullName(), "is already in our DB. Do you want to overwrite it? (y/n)");
+
+                        if(input.next().toLowerCase().startsWith("y")){
+                            System.out.println("Overwriting....");
+                            contactUpdate(fullContactName);
+                            System.out.println("Please enter new phone number: ");
+                            String userNewNumber = input.next();
+                            List<String> lines = Files.readAllLines(path, StandardCharsets.UTF_8);
+                            int position = 3;
+                            String extraline = String.format("%20s(\"%s %s\",\"%s\"),", "new Contacts", userNewFirst, userNewLast, userNewNumber);
+                            System.out.println(extraline);
+                            lines.add(position, extraline);
+                            Files.write(path, lines, StandardCharsets.UTF_8);
+                            openMenu();
+                        }
+                        else openMenu();
+                    }
+                }
+                    System.out.println("Please enter new phone number: ");
+                    String userNewNumber = input.next();
+                    List<String> lines = Files.readAllLines(path, StandardCharsets.UTF_8);
+                    int position = 3;
+                    String extraline = String.format("%20s(\"%s %s\",\"%s\"),", "new Contacts", userNewFirst, userNewLast, userNewNumber);
+                    System.out.println(extraline);
+                    lines.add(position, extraline);
+                    Files.write(path, lines, StandardCharsets.UTF_8);
+                    System.exit(0);
             }
         } catch (IOException IOException) {
             System.err.println("Exception " + IOException);
@@ -95,13 +118,17 @@ public class contactsManagerTest {
 
     public static void contactSearch(){
         Scanner input = new Scanner(System.in);
-        System.out.println("Please enter a name to search: ");
+        System.out.println("Please enter a full name to search: ");
         String userName = input.nextLine();
         for(Contacts contact: contactNew)
             if (contact.getFullName().equalsIgnoreCase(userName)){
                 System.out.printf("%-20s   |   %s\n\n",contact.getFullName(),contact.getPhoneNumber());
             }
     }
+
+//    public static void contactFind(String contactName){
+//
+//    }
 
     public static void contactDelete(){
         Scanner input = new Scanner(System.in);
@@ -145,4 +172,42 @@ public class contactsManagerTest {
             System.err.println("Exception " + IOException);
         }
     }
-} // EOF -> End of file
+ // EOF -> End of file
+
+    public static void contactUpdate(String updateContact){
+        String line;
+        String userInput = "";
+        try {
+            Path path = Paths.get("src/ContactsArray.java");
+            BufferedReader file = new BufferedReader(new FileReader("src/ContactsArray.java"));
+            if (!Files.exists(path)) {
+                System.out.println("Cannot execute your request, file doesn't exist");
+            } else if (Files.exists(path)) {
+                for(Contacts contact: contactNew)
+                    if (contact.getFullName().equalsIgnoreCase(updateContact)){
+                        System.out.printf("%-20s   |   %s\n\n",contact.getFullName(),contact.getPhoneNumber());
+
+                        List<String> lines = Files.readAllLines(path, StandardCharsets.UTF_8);
+                        String lineToDelete = String.format("%20s(\"%s\",\"%s\"),", "new Contacts", updateContact, contact.getPhoneNumber());
+
+                        while ((line = file.readLine()) != null)
+                        {
+                            //System.out.println(line);
+                            if (line.contains(lineToDelete))
+                            {
+                                line = "";
+                            }
+                            userInput += line + '\n';
+                        }
+                        FileOutputStream File = new FileOutputStream("src/ContactsArray.java");
+                        File.write(userInput.getBytes());
+                        file.close();
+                        File.close();
+                    }
+            }
+        } catch (IOException IOException) {
+            System.err.println("Exception " + IOException);
+        }
+    }
+}
+
